requirements.txt
langchain
langchain-core
langchain-ollama
langgraph
python-dotenv
